fix(apiserver): reconcile secs_to_rfc3339 call after merging main (PR #712)
PR #712 changed secs_to_rfc3339 to take i64 directly (removing a lossy
`as u64` cast that mishandled pre-1970 dates). This branch's CRD status
decode still passed `secs as u64`, breaking the build after merging main.

+++ crates/apiserver/src/apiextensions_gen_adapter.rs
@@ -685,9 +685,8 @@ pub fn decode_crd_proto_gen(data: &[u8]) -> Option<serde_json::Value> {
                     }
                     if let Some(t) = c.last_transition_time {
                         if let Some(secs) = t.seconds.filter(|&s| s > 0) {
-                            cm["lastTransitionTime"] = serde_json::Value::String(
-                                crate::util::secs_to_rfc3339(secs as u64),
-                            );
+                            cm["lastTransitionTime"] =
+                                serde_json::Value::String(crate::util::secs_to_rfc3339(secs));
                         }
                     }
                     cm
